+++ nio/src/main/scala/zio/nio/file/FileSystem.scala
@@ -1,18 +1,19 @@
-package zio.nio.core.file
+package zio.nio.file
 
 import java.net.URI
 import java.nio.file.attribute.UserPrincipalLookupService
 import java.nio.{ file => jf }
 
 import zio.blocking.{ Blocking, effectBlocking }
-import zio.{ UIO, ZIO }
+import zio.nio.core.file.{ Path, WatchService }
+import zio.{ UIO, ZIO, ZManaged }
 
 import scala.jdk.CollectionConverters._
 
 final class FileSystem private (private val javaFileSystem: jf.FileSystem) {
   def provider: jf.spi.FileSystemProvider = javaFileSystem.provider()
 
-  def close: ZIO[Blocking, Exception, Unit] =
+  private def close: ZIO[Blocking, Exception, Unit] =
     effectBlocking(javaFileSystem.close()).refineToOrDie[Exception]
 
   def isOpen: UIO[Boolean] = UIO.effectTotal(javaFileSystem.isOpen())
@@ -38,22 +39,27 @@ final class FileSystem private (private val javaFileSystem: jf.FileSystem) {
 }
 
 object FileSystem {
+  private val close: FileSystem => ZIO[Blocking, Nothing, Unit] = _.close.orDie
+
   def fromJava(javaFileSystem: jf.FileSystem): FileSystem = new FileSystem(javaFileSystem)
 
   def default: FileSystem = new FileSystem(jf.FileSystems.getDefault)
 
-  def getFileSystem(uri: URI): ZIO[Blocking, Exception, FileSystem] =
-    effectBlocking(new FileSystem(jf.FileSystems.getFileSystem(uri))).refineToOrDie[Exception]
+  def getFileSystem(uri: URI): ZManaged[Blocking, Exception, FileSystem] =
+    effectBlocking(new FileSystem(jf.FileSystems.getFileSystem(uri))).refineToOrDie[Exception].toManaged(close)
 
-  def newFileSystem(uri: URI, env: (String, Any)*): ZIO[Blocking, Exception, FileSystem] =
+  def newFileSystem(uri: URI, env: (String, Any)*): ZManaged[Blocking, Exception, FileSystem] =
     effectBlocking(new FileSystem(jf.FileSystems.newFileSystem(uri, env.toMap.asJava)))
       .refineToOrDie[Exception]
+      .toManaged(close)
 
-  def newFileSystem(uri: URI, env: Map[String, _], loader: ClassLoader): ZIO[Blocking, Exception, FileSystem] =
+  def newFileSystem(uri: URI, env: Map[String, _], loader: ClassLoader): ZManaged[Blocking, Exception, FileSystem] =
     effectBlocking(new FileSystem(jf.FileSystems.newFileSystem(uri, env.asJava, loader)))
       .refineToOrDie[Exception]
+      .toManaged(close)
 
-  def newFileSystem(path: Path, loader: ClassLoader): ZIO[Blocking, Exception, FileSystem] =
+  def newFileSystem(path: Path, loader: ClassLoader): ZManaged[Blocking, Exception, FileSystem] =
     effectBlocking(new FileSystem(jf.FileSystems.newFileSystem(path.javaPath, loader)))
       .refineToOrDie[Exception]
+      .toManaged(close)
 }
